Move CAPI error utility to 'errors' file
