+++ internal/errors/errors_test.go
@@ -1,4 +1,4 @@
-package key
+package errors
 
 import (
 	"errors"
