+++ pkg/azurecluster/validate_update.go
@@ -14,7 +14,6 @@ import (
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
 	"github.com/giantswarm/azure-admission-controller/internal/semverhelper"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
@@ -55,7 +54,7 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	err := azureClusterNewCR.ValidateUpdate(azureClusterOldCR)
-	err = key.IgnoreCAPIErrorForField("metadata.Name", err)
+	err = errors.IgnoreCAPIErrorForField("metadata.Name", err)
 	if err != nil {
 		return microerror.Mask(err)
 	}
