+++ pkg/azuremachine/validate_create.go
@@ -12,7 +12,6 @@ import (
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
@@ -61,7 +60,7 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	err := cr.ValidateCreate()
-	err = key.IgnoreCAPIErrorForField("sshPublicKey", err)
+	err = errors.IgnoreCAPIErrorForField("sshPublicKey", err)
 	if err != nil {
 		return microerror.Mask(err)
 	}
