+++ pkg/azuremachine/validate_update.go
@@ -13,7 +13,6 @@ import (
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
 	"github.com/giantswarm/azure-admission-controller/internal/semverhelper"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
@@ -54,7 +53,7 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	err := azureMachineNewCR.ValidateUpdate(azureMachineOldCR)
-	err = key.IgnoreCAPIErrorForField("sshPublicKey", err)
+	err = errors.IgnoreCAPIErrorForField("sshPublicKey", err)
 	if err != nil {
 		return microerror.Mask(err)
 	}
